Generate `superuser_stats` ZIP asynchronously
Fixes #340
